create: controllers and routes

+++ src/configurations/rest_errors/rest_error.go
@@ -7,7 +7,7 @@ import (
 type RestError struct {
 	Message string   `json:"message"`
 	Err     string   `json:"error"`
-	Code    int64    `json:"code"`
+	Code    int      `json:"code"`
 	Causes  []Causes `json:"causes"`
 }
 
@@ -20,7 +20,7 @@ type Causes struct {
 	Message string `json:"message"`
 }
 
-func NewRestErr(message, err string, code int64, causes []Causes) *RestError {
+func NewRestErr(message, err string, code int, causes []Causes) *RestError {
 	return &RestError{
 		Message: message,
 		Err:     err,
